bugfix for tcp and udp samplers
There was a bug in the `nested_map_from_file()` function. The inner
`while` should have been an `if` so that the pkey is refreshed
with each iteration.

This change moves the function into `crate::common` and fixes the
bug.

Additionally, we fix sampler intialization so that perf and bpf
are not initialized when the sampler is disabled.
