+++ src/samplers/udp/mod.rs
@@ -2,14 +2,10 @@
 // Licensed under the Apache License, Version 2.0
 // http://www.apache.org/licenses/LICENSE-2.0
 
-use std::collections::HashMap;
-use std::path::Path;
 use std::sync::Arc;
 
 use async_trait::async_trait;
 use metrics::*;
-use tokio::fs::File;
-use tokio::io::{AsyncBufReadExt, BufReader};
 use tokio::runtime::Handle;
 
 use crate::config::{Config, SamplerConfig};
@@ -72,7 +68,7 @@ impl Sampler for Udp {
         self.register();
 
         // sample /proc/net/snmp
-        if let Ok(snmp) = nested_map_from_file("/proc/net/snmp").await {
+        if let Ok(snmp) = crate::common::nested_map_from_file("/proc/net/snmp").await {
             let time = time::precise_time_ns();
             for statistic in self.sampler_config().statistics() {
                 if let Some((pkey, lkey)) = statistic.keys() {
@@ -86,7 +82,7 @@ impl Sampler for Udp {
         }
 
         // sample /proc/net/netstat
-        if let Ok(snmp) = nested_map_from_file("/proc/net/netstat").await {
+        if let Ok(snmp) = crate::common::nested_map_from_file("/proc/net/netstat").await {
             let time = time::precise_time_ns();
             for statistic in self.sampler_config().statistics() {
                 if let Some((pkey, lkey)) = statistic.keys() {
@@ -113,31 +109,4 @@ impl Sampler for Udp {
             Some(self.general_config().window()),
         ))
     }
-}
-
-async fn nested_map_from_file<T: AsRef<Path>>(
-    path: T,
-) -> Result<HashMap<String, HashMap<String, u64>>, std::io::Error> {
-    let mut ret = HashMap::<String, HashMap<String, u64>>::new();
-    let file = File::open(path).await?;
-    let reader = BufReader::new(file);
-    let mut lines = reader.lines();
-    while let Some(keys) = lines.next_line().await? {
-        while let Some(values) = lines.next_line().await? {
-            let keys: Vec<&str> = keys.trim().split_whitespace().collect();
-            let values: Vec<&str> = values.trim().split_whitespace().collect();
-            if keys.len() > 2 {
-                let pkey = keys[0];
-                if !ret.contains_key(pkey) {
-                    ret.insert(pkey.to_string(), Default::default());
-                }
-                let inner = ret.get_mut(&pkey.to_string()).unwrap();
-                for (i, key) in keys.iter().enumerate().skip(1) {
-                    let value: u64 = values.get(i).unwrap_or(&"0").parse().unwrap_or(0);
-                    inner.insert((*key).to_string(), value);
-                }
-            }
-        }
-    }
-    Ok(ret)
 }
